pyproject.toml
[tool.poetry]
name = "Whitelist Censor Service Client"
version = "0.1.0"
description = ""
authors = ["RRR-09 <79768142+RRR-09@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
fastapi = "^0.83.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
aiohttp = "^3.8.1"
websockets = "^10.3"
paramiko = "^2.11.0"
rocketry = "^2.4.1"
twitchio = "^2.5.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
bandit = "^1.7.4"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.971"
ossaudit = "^0.5.0"
types-paramiko = "^2.11.6"
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
src_paths=["src"]
profile = "black"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
show_error_codes = true
